Rename Menus to `MenuDesktop` and `MenuMobile`, Add custom `MenuMobile` styling
