+++ src/10_styles/02_molecule-styles/_m-menu-mobile.scss
@@ -1,8 +1,8 @@
 /*
- * Burger Menu Styles
+ * Mobile Menu Styles
  *
  * Custom styling for the Reach UI 'menu-button' component: https://reacttraining.com/reach-ui/menu-button
- * Source code: https://github.com/reach/reach-ui/tree/master/packages/skip-nav
+ * Source code: https://github.com/reach/reach-ui/tree/main/packages/menu-button
  */
 
 [data-reach-menu-button] {
@@ -17,6 +17,17 @@
     &:focus {
         background-color: $button-background--hover;
     }
+
+    @media (min-width: $breakpoint-tablet) {
+        display: none;
+    }
+}
+
+[data-reach-menu-popover] {
+    left: auto !important; // Overrides default so that menu isn't off screen
+    right: $gutter--m !important; // Overrides default so that menu isn't off screen
+    margin-left: $gutter--m;
+    width: auto;
 }
 
 [data-reach-menu-list] {
@@ -46,4 +57,19 @@
         color: $button-color;
         background: $button-background;
     }
+}
+
+@keyframes slide-down {
+    0% {
+        opacity: 0;
+        transform: translateY(-20px);
+    }
+    100% {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+.slide-down[data-reach-menu-list],
+.slide-down[data-reach-menu-items] {
+    animation: slide-down 0.2s ease;
 }
