+++ src/10_styles/styles.scss
@@ -23,8 +23,8 @@
 @import "01_atom-styles/a-ex-link";
 
 // Molecule Styles
-@import "02_molecule-styles/m-nav-links";
-@import "02_molecule-styles/m-menu-burger";
+@import "02_molecule-styles/m-menu-mobile";
+@import "02_molecule-styles/m-menu-desktop";
 
 // Organism Styles
 @import "03_organism-styles/o-header";
